Add CompressionStream compression (#126)
The compression stream uses gzip which is, as of now, the new default for the visualizer.
The main point is to reduce external dependencies, especially when a simpler way exists with a better result (as of benchmarking)

Decompression of URLs that contain lz-string compressed data is still supported for the forseeable future, but may be removed. The dependency is loaded on demand.